Create a CI job to release the add-on
The CI job will take care of all the steps to create a new release in
GitLab:
* creation of the tag
* create of the release object with the release notes
All the required information will be extracted from the addon.xml file.

A "pre-release" job is also added to validate the changes in addon.xml
before the actual release is done.

All these steps are explained in the contribution guidelines.

Finally the files TESTME.md and createaddon.sh are removed since the
installation steps are explained in the wiki and the archive of the
add-on is created automatically by GitLab in the release.

See merge request StCyr/plugin.video.peertube!12 for more information