Create a CI job to check strings.po files
The new "translation" job that will be available only when strings.po
files are modified.
It will use the msgcmp tool to check that the translation files use the
correct reference strings.

A new part is also added in the contribution guidelines to help future
translators to start up.

See merge request StCyr/plugin.video.peertube!21 for more information