Download torrents with vfs.libtorrent
Replace python-libtorrent with vfs.libtorrent.
Some parts of the code were removed as they will probably be useless
in the future.
The service is disabled (temporarily?) because currently the playback starts
only when the whole file is downloaded (no background download anymore).

Note that the path of the downloaded file is built manually because
vfs.libtorrent does not return this information yet.
The extension of the file is forced to "mp4" so it may not work for
all the videos.