Replace python-libtorrent with vfs.libtorrent
Now torrents are downloaded with vfs.libtorrent[1].

Some parts of the code were removed as they will probably be useless
in the future. For instance, the service is disabled (temporarily?)
because the download in background/foreground is now managed by
vfs.libtorrent.

Use the Read() API of vfs.libtorrent to retrieve the path where the
torrent was downloaded.
Also:
* Display a warning if there are more than 1 file in the torrent
because the add-on will play only the first file
* Remove the "start downloading" notification since an equivalent
  message is now displayed by vfs.libtorrent

[1]:https://framagit.org/thombet/vfs.libtorrent