Pause download when playback stops
Use the add-on's service to monitor when the playback is stopped.
When it occurs, the download of the torrent is paused to have a
behavior similar to streaming.

To be able to pause the torrent, the service needs to use the
torrent handle. The torrent handle is recreated on the service-side
using the URL that was sent to vfs.libtorrent. This URL is sent to
the service before the playback starts using the AddonSignals.
Calling `xbmcvfs.File()` on existing torrent is fast enough for this
solution to be used even on slow devices.