Don't try to play again the video in case of error
When we detected a playback error due to a too small downloaded file
the user had the choice to try again to play the video. On OSMC+raspi3
it was leading to a crash everytime the user was trying to play again
the video.
To avoid this now the user is only offered the option to pause or not
the download: the user will have to manually play again the video.
Depending on the user feedbacks and the test on other platforms (for
instance if the code that was removed is actually working on other
platforms), we may enable back this feature.